rename: 모듈이름, 패키지 이름 변경
