--- api/src/main/java/org/example/Main.java
@@ -1,7 +1,0 @@
-package org.example;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-}
