--- api/src/main/java/org/example/config/PropertiesConfig.java
@@ -1,9 +1,0 @@
-package org.example.config;
-
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationPropertiesScan
-public class PropertiesConfig {
-}
