+++ app/build.gradle
@@ -40,7 +40,7 @@ java {
 
 application {
     // Define the main class for the application.
-    mainClass = 'org.example.App'
+    mainClass = 'org.healthmap.app.App'
 }
 
 tasks.named('test') {
