--- app/src/test/java/org/example/AppTest.java
@@ -1,12 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.example;
-
-import org.junit.jupiter.api.Test;
-
-class AppTest {
-    @Test
-    void appHasAGreeting() {
-    }
-}
