--- app/src/test/java/org/example/controller/DrugstoreControllerTest.java
@@ -1,52 +1,0 @@
-package org.example.controller;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-class DrugstoreControllerTest {
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private WebApplicationContext webApplicationContext;
-
-    String baseUrl = "/drugstore";
-
-    @BeforeEach
-    public void setup() {
-        this.mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-                .addFilter(new CharacterEncodingFilter("UTF-8", true))
-                .alwaysDo(MockMvcResultHandlers.print())
-                .build();
-    }
-
-    @Test
-    @DisplayName("전체 약국 가져오는지 확인")
-    public void testFindAllDrugstore() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.get(baseUrl))
-                .andExpect(MockMvcResultMatchers.status().isOk())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Lee"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.address").value("서울시"));
-    }
-
-    @Test
-    @DisplayName("사용자 근처의 약국 가져오는지 확인")
-    public void testFindDrugstoreAround() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.get(baseUrl + "/around?longitude=123.11111&latitude=53.222222"))
-                .andExpect(MockMvcResultMatchers.status().isOk())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.length()").value("2"));
-    }
-
-}
