+++ app/src/test/java/org/healthmap/AppTest.java
@@ -1,0 +1,12 @@
+/*
+ * This source file was generated by the Gradle 'init' task
+ */
+package org.healthmap;
+
+import org.junit.jupiter.api.Test;
+
+class AppTest {
+    @Test
+    void appHasAGreeting() {
+    }
+}
