+++ openapi/src/main/java/org/healthmap/OpenApiApplication.java
@@ -1,0 +1,11 @@
+package org.healthmap;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OpenApiApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(OpenApiApplication.class, args);
+    }
+}
