+++ openapi/src/main/java/org/healthmap/openapi/config/PropertiesConfig.java
@@ -1,0 +1,9 @@
+package org.healthmap.openapi.config;
+
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationPropertiesScan
+public class PropertiesConfig {
+}
